Add menu

+++ app/src/main/java/me/annenkov/julistaandroid/data/model/booklet/auth/Students.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Students(
         @SerializedName("default")
-        val jsonMemberDefault: Int? = null,
+        val jsonMemberDefault: Long? = null,
 
         @SerializedName("list")
         val list: List<Student>? = null
